+++ src/js/mixin/internal/animate-fade.js
@@ -118,16 +118,7 @@ function waitTransitionend(target) {
 }
 
 function getTransitionNodes(target) {
-    return getRows(children(target)).reduce(
-        (nodes, row) =>
-            nodes.concat(
-                sortBy(
-                    row.filter((el) => isInView(el)),
-                    'offsetLeft'
-                )
-            ),
-        []
-    );
+    return getRows(children(target)).flat();
 }
 
 function awaitFrame() {